Prototype/requirements.txt
markdown
flask
pandas
scikit-learn
numpy
flask-bcrypt
Flask-JWT-Extended
python-dotenv
langchain 
langchain-community 
langchain_google_genai
streamlit 